Fixed integer precision loss warning in test support class.
	Change on 2016/08/31 by tball <tball@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=131861717

+++ jre_emul/Tests/com/google/j2objc/NativeUtil.java
@@ -22,6 +22,6 @@ package com.google.j2objc;
 public class NativeUtil {
 
   public static native int getRetainCount(Object o) /*-[
-    return ((NSObject *)o).retainCount;
+    return (jint) ((NSObject *)o).retainCount;
   ]-*/;
 }
